[CHORE] UserLinkService 추상화
